Add patient related DAOs
