+++ app/src/main/java/org/resolvetosavelives/red/newentry/search/Patient.kt
@@ -1,9 +1,14 @@
 package org.resolvetosavelives.red.newentry.search
 
+import android.arch.persistence.room.Dao
 import android.arch.persistence.room.Entity
 import android.arch.persistence.room.ForeignKey
 import android.arch.persistence.room.Index
+import android.arch.persistence.room.Insert
+import android.arch.persistence.room.OnConflictStrategy
 import android.arch.persistence.room.PrimaryKey
+import android.arch.persistence.room.Query
+import io.reactivex.Flowable
 import org.threeten.bp.Instant
 import org.threeten.bp.LocalDate
 
@@ -42,4 +47,18 @@ data class Patient(
     val updatedAt: Instant,
 
     val syncPending: Boolean
-)
+) {
+
+  @Dao
+  interface RoomDao {
+
+    @Query("SELECT * FROM patient WHERE fullName LIKE '%' || :query || '%'")
+    fun search(query: String): Flowable<List<Patient>>
+
+    @Query("SELECT * FROM patient")
+    fun allPatients(): Flowable<List<Patient>>
+
+    @Insert(onConflict = OnConflictStrategy.REPLACE)
+    fun save(patient: Patient)
+  }
+}
