+++ app/src/main/java/org/resolvetosavelives/red/newentry/search/PatientAddress.kt
@@ -1,6 +1,9 @@
 package org.resolvetosavelives.red.newentry.search
 
+import android.arch.persistence.room.Dao
 import android.arch.persistence.room.Entity
+import android.arch.persistence.room.Insert
+import android.arch.persistence.room.OnConflictStrategy
 import android.arch.persistence.room.PrimaryKey
 import org.threeten.bp.Instant
 
@@ -22,4 +25,12 @@ data class PatientAddress(
     val updatedAt: Instant,
 
     val syncPending: Boolean
-)
+) {
+
+  @Dao
+  interface RoomDao {
+
+    @Insert(onConflict = OnConflictStrategy.REPLACE)
+    fun save(address: PatientAddress)
+  }
+}
