+++ app/src/main/java/org/resolvetosavelives/red/newentry/search/PatientPhoneNumber.kt
@@ -1,6 +1,9 @@
 package org.resolvetosavelives.red.newentry.search
 
+import android.arch.persistence.room.Dao
 import android.arch.persistence.room.Entity
+import android.arch.persistence.room.Insert
+import android.arch.persistence.room.OnConflictStrategy
 import android.arch.persistence.room.PrimaryKey
 import org.threeten.bp.Instant
 
@@ -20,4 +23,12 @@ data class PatientPhoneNumber(
     val updatedAt: Instant,
 
     val syncPending: Boolean
-)
+) {
+
+  @Dao
+  interface RoomDao {
+
+    @Insert(onConflict = OnConflictStrategy.REPLACE)
+    fun save(phoneNumber: PatientPhoneNumber)
+  }
+}
